requirements.all.txt
django-ace>=1.0.1
django-classy-tags>=0.4
django-haystack>=2.0,<3
django-imagekit>=3.0.4
django-model-utils>=2.0
django-tinymce>=1.5.2
django-wymeditor>=1.0
-e git+git@github.com:kezabelle/django-setuptest.git@feature/django17_support#egg=django-setuptest
-e git+https://github.com/kezabelle/django-adminlinks.git#egg=django-adminlinks
feedparser>=5.1.3,<6.0
Pillow>=2.3
PyYAML>=3.10
South>=0.8
toml>=0.8.1
